movies_api/requirements.txt
redis==4.6.0
elasticsearch[async]==7.9.1
fastapi==0.100.0
orjson==3.9.2
pydantic==1.10.9
uvicorn==0.22.0
gunicorn==20.1.0
httptools==0.6.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
grpcio==1.57.0
grpcio-tools==1.57.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.19.0